requirements.txt
Pillow
numpy
svgpathtools
matplotlib

